gdb, refactor into module
